+++ CHANGELOG.md
@@ -31,7 +31,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Fixed
 
-- 誤字を修正
+- いくつかの軽微な修正
 
 ## [1.5.1] - 2022-05-31